initial dashboard commit for  task-5
